seeder/requirements.txt
pymongo
python-dotenv
numpy
